【重构】重构IExportFileByTemplate中的ExportByTemplate，将参数htmlTemplate改为temp…
…late。以便支持Excel模板导出。

- 【导出】支持Excel模板导出并填写相关单元测试，如何使用见教程《Excel模板导出之导出教材订购表》
    - 支持单元格单个绑定
    - 支持列表

+++ 01.Skynet.Cloud.Framework/2.UnitTest/Skynet.Cloud.IE.Test/ExcelImporter_Tests.cs
@@ -1,8 +1,5 @@
 // ======================================================================
 // 
-//           Copyright (C) 2019-2030 ÉîÛÚÊĞÓÅÍø¿Æ¼¼ÓĞÏŞ¹«Ë¾
-//           All rights reserved
-// 
 //           filename : ExcelImporter_Tests.cs
 //           description :
 // 
@@ -10,51 +7,72 @@
 //           
 //           
 //           
-//          
+//           
 // 
 // ======================================================================
 
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Threading.Tasks;
 using UWay.Skynet.Cloud.IE.Core;
 using UWay.Skynet.Cloud.IE.Core.Extension;
 using UWay.Skynet.Cloud.IE.Core.Models;
 using UWay.Skynet.Cloud.IE.Excel;
-using UWay.Skynet.Cloud.IE.Tests.Models;
+using UWay.Skynet.Cloud.IE.Tests.Models.Import;
+using Newtonsoft.Json;
 using Shouldly;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
 using Xunit;
+using Xunit.Abstractions;
 
 namespace UWay.Skynet.Cloud.IE.Tests
 {
-    public class ExcelImporter_Tests
+    public class ExcelImporter_Tests : TestBase
     {
+        public ExcelImporter_Tests(ITestOutputHelper testOutputHelper)
+        {
+            _testOutputHelper = testOutputHelper;
+        }
+
+        private readonly ITestOutputHelper _testOutputHelper;
         public IImporter Importer = new ExcelImporter();
 
-        [Fact(DisplayName = "Éú³ÉÄ£°å")]
+        /// <summary>
+        /// æµ‹è¯•æšä¸¾
+        /// </summary>
+        /// <returns></returns>
+        [Fact(DisplayName = "ç”Ÿæˆå­¦ç”Ÿæ•°æ®å¯¼å…¥æ¨¡æ¿")]
+        public async Task GenerateStudentImportTemplate_Test()
+        {
+            var filePath = Path.Combine(Directory.GetCurrentDirectory(),
+                nameof(GenerateStudentImportTemplate_Test) + ".xlsx");
+            if (File.Exists(filePath)) File.Delete(filePath);
+
+            var result = await Importer.GenerateTemplate<ImportStudentDto>(filePath);
+            result.ShouldNotBeNull();
+            File.Exists(filePath).ShouldBeTrue();
+
+            //TODO:è¯»å–Excelæ£€æŸ¥è¡¨å¤´å’Œæ ¼å¼
+        }
+
+        [Fact(DisplayName = "ç”Ÿæˆæ¨¡æ¿")]
         public async Task GenerateTemplate_Test()
         {
             var filePath = Path.Combine(Directory.GetCurrentDirectory(), nameof(GenerateTemplate_Test) + ".xlsx");
-            if (File.Exists(filePath))
-            {
-                File.Delete(filePath);
-            }
+            if (File.Exists(filePath)) File.Delete(filePath);
 
             var result = await Importer.GenerateTemplate<ImportProductDto>(filePath);
             result.ShouldNotBeNull();
             File.Exists(filePath).ShouldBeTrue();
 
-            //TODO:ÁĞÍ·Òş²Ø²âÊÔ
+            //TODO:è¯»å–Excelæ£€æŸ¥è¡¨å¤´å’Œæ ¼å¼
         }
 
-        [Fact(DisplayName = "Éú³ÉÄ£°å×Ö½Ú")]
+        [Fact(DisplayName = "ç”Ÿæˆæ¨¡æ¿å­—èŠ‚")]
         public async Task GenerateTemplateBytes_Test()
         {
-            var filePath = Path.Combine(Directory.GetCurrentDirectory(), nameof(GenerateTemplateBytes_Test) +".xlsx");
-            if (File.Exists(filePath))
-            {
-                File.Delete(filePath);
-            }
+            var filePath = Path.Combine(Directory.GetCurrentDirectory(), nameof(GenerateTemplateBytes_Test) + ".xlsx");
+            if (File.Exists(filePath)) File.Delete(filePath);
 
             var result = await Importer.GenerateTemplateBytes<ImportProductDto>();
             result.ShouldNotBeNull();
@@ -63,12 +81,18 @@ namespace UWay.Skynet.Cloud.IE.Tests
             File.Exists(filePath).ShouldBeTrue();
         }
 
-        [Fact(DisplayName = "µ¼Èë")]
+        /// <summary>
+        /// æµ‹è¯•ï¼š
+        /// è¡¨å¤´è¡Œä½ç½®è®¾ç½®
+        /// å¯¼å…¥é€»è¾‘æµ‹è¯•
+        /// </summary>
+        /// <returns></returns>
+        [Fact(DisplayName = "äº§å“ä¿¡æ¯å¯¼å…¥")]
         public async Task Importer_Test()
         {
-            //µÚÒ»ÁĞÂÒĞò
+            //ç¬¬ä¸€åˆ—ä¹±åº
 
-            var filePath = Path.Combine(Directory.GetCurrentDirectory(), "TestFiles", "Import", "²úÆ·µ¼ÈëÄ£°å.xlsx");
+            var filePath = Path.Combine(Directory.GetCurrentDirectory(), "TestFiles", "Import", "äº§å“å¯¼å…¥æ¨¡æ¿.xlsx");
             var import = await Importer.Import<ImportProductDto>(filePath);
             import.ShouldNotBeNull();
 
@@ -77,40 +101,54 @@ namespace UWay.Skynet.Cloud.IE.Tests
             import.Data.Count.ShouldBeGreaterThanOrEqualTo(2);
             foreach (var item in import.Data)
             {
-                if (item.Name.Contains("¿Õ¸ñ²âÊÔ"))
-                {
-                    item.Name.ShouldBe(item.Name.Trim());
-                }
+                if (item.Name.Contains("ç©ºæ ¼æµ‹è¯•")) item.Name.ShouldBe(item.Name.Trim());
 
-                if (item.Code.Contains("²»È¥³ı¿Õ¸ñ²âÊÔ"))
-                {
-                    item.Code.ShouldContain(" ");
-                }
-                //È¥³ıÖĞ¼ä¿Õ¸ñ²âÊÔ
+                if (item.Code.Contains("ä¸å»é™¤ç©ºæ ¼æµ‹è¯•")) item.Code.ShouldContain(" ");
+                //å»é™¤ä¸­é—´ç©ºæ ¼æµ‹è¯•
                 item.BarCode.ShouldBe("123123");
             }
 
-            //¿ÉÎª¿ÕÀàĞÍ²âÊÔ
+            //å¯ä¸ºç©ºç±»å‹æµ‹è¯•
             import.Data.ElementAt(4).Weight.HasValue.ShouldBe(true);
             import.Data.ElementAt(5).Weight.HasValue.ShouldBe(false);
-            //ÌáÈ¡ĞÔ±ğ¹«Ê½²âÊÔ
-            import.Data.ElementAt(0).Sex.ShouldBe("Å®");
-            //»ñÈ¡µ±Ç°ÈÕÆÚÒÔ¼°ÈÕÆÚÀàĞÍ²âÊÔ  Èç¹ûÊ±¼ä²»¶Ô£¬Çë´ò¿ª¶ÔÓ¦µÄExcel¼´¿É¸üĞÂÎªµ±Ç°Ê±¼ä£¬È»ºóÔÙÔËĞĞ´Ëµ¥Ôª²âÊÔ
+            //æå–æ€§åˆ«å…¬å¼æµ‹è¯•
+            import.Data.ElementAt(0).Sex.ShouldBe("å¥³");
+            //è·å–å½“å‰æ—¥æœŸä»¥åŠæ—¥æœŸç±»å‹æµ‹è¯•  å¦‚æœæ—¶é—´ä¸å¯¹ï¼Œè¯·æ‰“å¼€å¯¹åº”çš„Excelå³å¯æ›´æ–°ä¸ºå½“å‰æ—¶é—´ï¼Œç„¶åå†è¿è¡Œæ­¤å•å…ƒæµ‹è¯•
             //import.Data[0].FormulaTest.Date.ShouldBe(DateTime.Now.Date);
-            //ÊıÖµ²âÊÔ
+            //æ•°å€¼æµ‹è¯•
             import.Data.ElementAt(0).DeclareValue.ShouldBe(123123);
             import.Data.ElementAt(0).Name.ShouldBe("1212");
             import.Data.ElementAt(0).BarCode.ShouldBe("123123");
             import.Data.ElementAt(1).Name.ShouldBe("12312312");
-            import.Data.ElementAt(2).Name.ShouldBe("×ó²à¿Õ¸ñ²âÊÔ");
+            import.Data.ElementAt(2).Name.ShouldBe("å·¦ä¾§ç©ºæ ¼æµ‹è¯•");
         }
 
-        [Fact(DisplayName = "±ØÌîÏî¼ì²â")]
+        [Fact(DisplayName = "æˆªæ–­æ•°æ®æµ‹è¯•")]
+        public async Task ImporterDataEnd_Test()
+        {
+            var filePath = Path.Combine(Directory.GetCurrentDirectory(), "TestFiles", "Import", "æˆªæ–­æ•°æ®æµ‹è¯•.xlsx");
+            var import = await Importer.Import<ImportProductDto2>(filePath);
+            import.ShouldNotBeNull();
+            import.Data.ShouldNotBeNull();
+            import.Data.Count.ShouldBe(6);
+        }
+
+        [Fact(DisplayName = "ç¼´è´¹æµæ°´å¯¼å…¥æµ‹è¯•")]
+        public async Task ImportPaymentLogs_Test()
+        {
+            var filePath = Path.Combine(Directory.GetCurrentDirectory(), "TestFiles", "Import", "ç¼´è´¹æµæ°´å¯¼å…¥æ¨¡æ¿.xlsx");
+            var import = await Importer.Import<ImportPaymentLogDto>(filePath);
+            import.ShouldNotBeNull();
+            import.HasError.ShouldBeTrue();
+            import.Exception.ShouldBeNull();
+            import.Data.Count.ShouldBe(20);
+        }
+
+        [Fact(DisplayName = "å¿…å¡«é¡¹æ£€æµ‹")]
         public async Task IsRequired_Test()
         {
             var pros = typeof(ImportProductDto).GetProperties();
             foreach (var item in pros)
-            {
                 switch (item.Name)
                 {
                     //DateTime
@@ -121,117 +159,133 @@ namespace UWay.Skynet.Cloud.IE.Tests
                     case "Name":
                         item.IsRequired().ShouldBe(true);
                         break;
-                    //¿ÉÎª¿ÕÀàĞÍ
+                    //å¯ä¸ºç©ºç±»å‹
                     case "Weight":
                     //string
                     case "IdNo":
                         item.IsRequired().ShouldBe(false);
                         break;
                 }
-            }
         }
 
-        [Fact(DisplayName = "Ìâ¿âµ¼Èë²âÊÔ")]
+        [Fact(DisplayName = "é¢˜åº“å¯¼å…¥æµ‹è¯•")]
         public async Task QuestionBankImporter_Test()
         {
-            var filePath = Path.Combine(Directory.GetCurrentDirectory(), "TestFiles", "Import", "Ìâ¿âµ¼ÈëÄ£°å.xlsx");
+            var filePath = Path.Combine(Directory.GetCurrentDirectory(), "TestFiles", "Import", "é¢˜åº“å¯¼å…¥æ¨¡æ¿.xlsx");
             var import = await Importer.Import<ImportQuestionBankDto>(filePath);
             import.ShouldNotBeNull();
-
+            _testOutputHelper.WriteLine(JsonConvert.SerializeObject(import.RowErrors));
             import.HasError.ShouldBeFalse();
             import.Data.ShouldNotBeNull();
             import.Data.Count.ShouldBe(404);
+
+            #region æ£€æŸ¥Boolå€¼æ˜ å°„
+
+            //æ˜¯
+            import.Data.ElementAt(0).IsDisorderly.ShouldBeTrue();
+            //å¦
+            import.Data.ElementAt(1).IsDisorderly.ShouldBeFalse();
+            //å¯¹
+            import.Data.ElementAt(2).IsDisorderly.ShouldBeTrue();
+            //é”™
+            import.Data.ElementAt(3).IsDisorderly.ShouldBeFalse();
+
+            #endregion
 
             import.RowErrors.Count.ShouldBe(0);
             import.TemplateErrors.Count.ShouldBe(0);
         }
 
-        [Fact(DisplayName = "½É·ÑÁ÷Ë®µ¼Èë²âÊÔ")]
-        public async Task ImportPaymentLogs_Test()
-        {
-            var filePath = Path.Combine(Directory.GetCurrentDirectory(), "TestFiles", "Import", "½É·ÑÁ÷Ë®µ¼ÈëÄ£°å.xlsx");
-            var import = await Importer.Import<ImportPaymentLogDto>(filePath);
-            import.ShouldNotBeNull();
-            import.HasError.ShouldBeTrue();
-            import.Exception.ShouldBeNull();
-            import.Data.Count.ShouldBe(20);
-        }
-
-        [Fact(DisplayName = "Êı¾İ´íÎó¼ì²â")]
+        [Fact(DisplayName = "æ•°æ®é”™è¯¯æ£€æµ‹")]
         public async Task RowDataError_Test()
         {
-            var filePath = Path.Combine(Directory.GetCurrentDirectory(), "TestFiles", "Errors", "Êı¾İ´íÎó.xlsx");
+            var filePath = Path.Combine(Directory.GetCurrentDirectory(), "TestFiles", "Errors", "æ•°æ®é”™è¯¯.xlsx");
             var result = await Importer.Import<ImportRowDataErrorDto>(filePath);
             result.ShouldNotBeNull();
             result.HasError.ShouldBeTrue();
 
             result.TemplateErrors.Count.ShouldBe(0);
 
-            result.RowErrors.ShouldContain(p => p.RowIndex == 2 && p.FieldErrors.ContainsKey("²úÆ·Ãû³Æ"));
-            result.RowErrors.ShouldContain(p => p.RowIndex == 3 && p.FieldErrors.ContainsKey("²úÆ·Ãû³Æ"));
+            result.RowErrors.ShouldContain(p => p.RowIndex == 2 && p.FieldErrors.ContainsKey("äº§å“åç§°"));
+            result.RowErrors.ShouldContain(p => p.RowIndex == 3 && p.FieldErrors.ContainsKey("äº§å“åç§°"));
 
-            result.RowErrors.ShouldContain(p => p.RowIndex == 7 && p.FieldErrors.ContainsKey("²úÆ·´úÂë"));
+            result.RowErrors.ShouldContain(p => p.RowIndex == 7 && p.FieldErrors.ContainsKey("äº§å“ä»£ç "));
 
-            result.RowErrors.ShouldContain(p => p.RowIndex == 3 && p.FieldErrors.ContainsKey("ÖØÁ¿(KG)"));
-            result.RowErrors.ShouldContain(p => p.RowIndex == 4 && p.FieldErrors.ContainsKey("¹«Ê½²âÊÔ"));
-            result.RowErrors.ShouldContain(p => p.RowIndex == 5 && p.FieldErrors.ContainsKey("¹«Ê½²âÊÔ"));
-            result.RowErrors.ShouldContain(p => p.RowIndex == 6 && p.FieldErrors.ContainsKey("¹«Ê½²âÊÔ"));
-            result.RowErrors.ShouldContain(p => p.RowIndex == 7 && p.FieldErrors.ContainsKey("¹«Ê½²âÊÔ"));
+            result.RowErrors.ShouldContain(p => p.RowIndex == 3 && p.FieldErrors.ContainsKey("é‡é‡(KG)"));
+            result.RowErrors.ShouldContain(p => p.RowIndex == 4 && p.FieldErrors.ContainsKey("å…¬å¼æµ‹è¯•"));
+            result.RowErrors.ShouldContain(p => p.RowIndex == 5 && p.FieldErrors.ContainsKey("å…¬å¼æµ‹è¯•"));
+            result.RowErrors.ShouldContain(p => p.RowIndex == 6 && p.FieldErrors.ContainsKey("å…¬å¼æµ‹è¯•"));
+            result.RowErrors.ShouldContain(p => p.RowIndex == 7 && p.FieldErrors.ContainsKey("å…¬å¼æµ‹è¯•"));
 
-            result.RowErrors.ShouldContain(p => p.RowIndex == 3 && p.FieldErrors.ContainsKey("Éí·İÖ¤"));
-            result.RowErrors.First(p => p.RowIndex == 3 && p.FieldErrors.ContainsKey("Éí·İÖ¤")).FieldErrors.Count
+            result.RowErrors.ShouldContain(p => p.RowIndex == 3 && p.FieldErrors.ContainsKey("èº«ä»½è¯"));
+            result.RowErrors.First(p => p.RowIndex == 3 && p.FieldErrors.ContainsKey("èº«ä»½è¯")).FieldErrors.Count
                 .ShouldBe(3);
 
-            result.RowErrors.ShouldContain(p => p.RowIndex == 4 && p.FieldErrors.ContainsKey("Éí·İÖ¤"));
-            result.RowErrors.ShouldContain(p => p.RowIndex == 5 && p.FieldErrors.ContainsKey("Éí·İÖ¤"));
+            result.RowErrors.ShouldContain(p => p.RowIndex == 4 && p.FieldErrors.ContainsKey("èº«ä»½è¯"));
+            result.RowErrors.ShouldContain(p => p.RowIndex == 5 && p.FieldErrors.ContainsKey("èº«ä»½è¯"));
 
-            #region ÖØ¸´´íÎó
+            #region é‡å¤é”™è¯¯
 
-            var errorRows = "5,6".Split(',').ToList();
+            var errorRows = new List<int>()
+            {
+                5,6
+            };
             result.RowErrors.ShouldContain(p =>
-                errorRows.Contains(p.RowIndex.ToString()) && p.FieldErrors.ContainsKey("²úÆ·´úÂë") &&
-                p.FieldErrors.Values.Contains("´æÔÚÊı¾İÖØ¸´£¬Çë¼ì²é£¡ËùÔÚĞĞ£º5£¬6¡£"));
+                errorRows.Contains(p.RowIndex) && p.FieldErrors.ContainsKey("äº§å“ä»£ç ") &&
+                p.FieldErrors.Values.Contains("å­˜åœ¨æ•°æ®é‡å¤ï¼Œè¯·æ£€æŸ¥ï¼æ‰€åœ¨è¡Œï¼š5ï¼Œ6ã€‚"));
 
-            errorRows = "8,9,11,13".Split(',').ToList();
+            errorRows = new List<int>()
+            {
+                8,9,11,13
+            };
             result.RowErrors.ShouldContain(p =>
-                errorRows.Contains(p.RowIndex.ToString()) && p.FieldErrors.ContainsKey("²úÆ·´úÂë") &&
-                p.FieldErrors.Values.Contains("´æÔÚÊı¾İÖØ¸´£¬Çë¼ì²é£¡ËùÔÚĞĞ£º8£¬9£¬11£¬13¡£"));
+                errorRows.Contains(p.RowIndex) && p.FieldErrors.ContainsKey("äº§å“ä»£ç ") &&
+                p.FieldErrors.Values.Contains("å­˜åœ¨æ•°æ®é‡å¤ï¼Œè¯·æ£€æŸ¥ï¼æ‰€åœ¨è¡Œï¼š8ï¼Œ9ï¼Œ11ï¼Œ13ã€‚"));
 
-            errorRows = "4£¬6£¬8£¬10£¬11£¬13".Split('£¬').ToList();
+            errorRows = new List<int>()
+            {
+                4,6,8,10,11,13
+            };
             result.RowErrors.ShouldContain(p =>
-                errorRows.Contains(p.RowIndex.ToString()) && p.FieldErrors.ContainsKey("²úÆ·ĞÍºÅ") &&
-                p.FieldErrors.Values.Contains("´æÔÚÊı¾İÖØ¸´£¬Çë¼ì²é£¡ËùÔÚĞĞ£º4£¬6£¬8£¬10£¬11£¬13¡£"));
+                errorRows.Contains(p.RowIndex) && p.FieldErrors.ContainsKey("äº§å“å‹å·") &&
+                p.FieldErrors.Values.Contains("å­˜åœ¨æ•°æ®é‡å¤ï¼Œè¯·æ£€æŸ¥ï¼æ‰€åœ¨è¡Œï¼š4ï¼Œ6ï¼Œ8ï¼Œ10ï¼Œ11ï¼Œ13ã€‚"));
+
             #endregion
 
             result.RowErrors.Count.ShouldBeGreaterThan(0);
 
-            //Ò»ĞĞ½öÔÊĞí´æÔÚÒ»ÌõÊı¾İ
+            //ä¸€è¡Œä»…å…è®¸å­˜åœ¨ä¸€æ¡æ•°æ®
             foreach (var item in result.RowErrors.GroupBy(p => p.RowIndex).Select(p => new { p.Key, Count = p.Count() }))
-            {
                 item.Count.ShouldBe(1);
-            }
+
+            char.Parse(",");
+            char.Parse("ï¼Œ");
         }
 
-        [Fact(DisplayName = "Ä£°å´íÎó¼ì²â")]
+        [Fact(DisplayName = "å­¦ç”ŸåŸºç¡€æ•°æ®å¯¼å…¥")]
+        public async Task StudentInfoImporter_Test()
+        {
+            var filePath = Path.Combine(Directory.GetCurrentDirectory(), "TestFiles", "Import", "å­¦ç”ŸåŸºç¡€æ•°æ®å¯¼å…¥.xlsx");
+            var import = await Importer.Import<ImportStudentDto>(filePath);
+            import.ShouldNotBeNull();
+            if (import.Exception != null) _testOutputHelper.WriteLine(import.Exception.ToString());
+
+            if (import.RowErrors.Count > 0) _testOutputHelper.WriteLine(JsonConvert.SerializeObject(import.RowErrors));
+            import.HasError.ShouldBeFalse();
+            import.Data.ShouldNotBeNull();
+            import.Data.Count.ShouldBe(16);
+        }
+
+        [Fact(DisplayName = "æ¨¡æ¿é”™è¯¯æ£€æµ‹")]
         public async Task TplError_Test()
         {
-            var filePath = Path.Combine(Directory.GetCurrentDirectory(), "TestFiles", "Errors", "Ä£°å×Ö¶Î´íÎó.xlsx");
-            var result = await Importer.Import<ImportProductDto>(filePath);
+            var filePath = Path.Combine(Directory.GetCurrentDirectory(), "TestFiles", "Errors", "æ¨¡æ¿å­—æ®µé”™è¯¯.xlsx");
+            var result = await Importer.Import<ImportProductDto2>(filePath);
             result.ShouldNotBeNull();
             result.HasError.ShouldBeTrue();
             result.TemplateErrors.Count.ShouldBeGreaterThan(0);
             result.TemplateErrors.Count(p => p.ErrorLevel == ErrorLevels.Error).ShouldBe(1);
             result.TemplateErrors.Count(p => p.ErrorLevel == ErrorLevels.Warning).ShouldBe(1);
-        }
-
-        [Fact(DisplayName = "½Ø¶ÏÊı¾İ²âÊÔ")]
-        public async Task ImporterDataEnd_Test()
-        {
-            var filePath = Path.Combine(Directory.GetCurrentDirectory(), "TestFiles", "Import", "½Ø¶ÏÊı¾İ²âÊÔ.xlsx");
-            var import = await Importer.Import<ImportProductDto>(filePath);
-            import.ShouldNotBeNull();
-            import.Data.ShouldNotBeNull();
-            import.Data.Count.ShouldBe(6);
         }
     }
 }
